@@ -4,7 +4,7 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
@@ -22,6 +22,9 @@ dependencies:
   popover: any
   http: any
   measured_size: any
+  sqflite: any
+  shared_preferences: any
+  uuid: any
 
 dev_dependencies:
   flutter_test:
