@@ -1,4 +1,5 @@
 const double BRIGHTNESS_THRESHOLD = 0.2;
-const double EDIT_DIALOG_MAX_WIDTH = 300;
+const double ALERT_DIALOG_MAX_WIDTH = 300;
+const double EDIT_DIALOG_MAX_WIDTH = 700;
 const double SEARCH_SELECT_DIALOG_MAX_HEIGHT = 500;
 const double SEARCH_SELECT_DIALOG_MAX_WIDTH = 400;
